Change type of "i" variables

+++ FillArrayLib/FillArrayLib.cpp
@@ -10,7 +10,7 @@ int diagonals_number(int columns, int rows)
 
 void fillArrayDown(std::array<std::array<int, maxColumns>, maxRows>& array, size_t rows, size_t columns, int start_row, int start_column, int& value)
 {
-	for (int i = 0; start_column - i >= 0 && start_row + i < rows; i++)
+	for (size_t i = 0; start_column - i >= 0 && start_row + i < rows; i++)
 	{
 		array[start_row + i][start_column - i] = value++;
 	}
@@ -18,7 +18,7 @@ void fillArrayDown(std::array<std::array<int, maxColumns>, maxRows>& array, size
 
 void fillArrayUp(std::array<std::array<int, maxColumns>, maxRows>& array, size_t rows, size_t columns, int start_row, int start_column, int& value)
 {
-	for (int i = 0; start_row - i >= 0 && start_column + i < columns; i++)
+	for (size_t i = 0; start_row - i >= 0 && start_column + i < columns; i++)
 	{
 		array[start_row - i][start_column + i] = value++;
 	}
